add crud role, delete user, update user

+++ PetSpa.Contract.Repositories/Entity/UserInfo.cs
@@ -5,7 +5,7 @@ namespace PetSpa.Contract.Repositories.Entity
     public class UserInfo:BaseEntity
     {
         public string FullName { get; set; } = string.Empty;
-        public double? Salary { get; set; }
+        public decimal? Salary { get; set; }
         public string? BankAccount { get; set; }
         public string? BankAccountName { get; set; }
         public string? Bank { get; set; }
